feat: "post to cancha"

+++ equipos/equipo-03/backend/src/main/java/com/canchavoley/backend/repository/CanchaRepository.java
@@ -3,6 +3,6 @@ package com.canchavoley.backend.repository;
 import com.canchavoley.backend.model.Cancha;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface CanchaRepository extends JpaRepository<Cancha, Long> {
+public interface CanchaRepository extends JpaRepository<Cancha, Integer> {
 
 }
